+++ btcstaking-tracker/stakingeventwatcher/stakingeventwatcher.go
@@ -71,6 +71,7 @@ type StakingEventWatcher struct {
 
 	btcClient   btcclient.BTCClient
 	btcNotifier notifier.ChainNotifier
+	indexer     SpendChecker
 	metrics     *metrics.UnbondingWatcherMetrics
 	// TODO: Ultimately all requests to babylon should go through some kind of semaphore
 	// to avoid spamming babylon with requests
@@ -95,6 +96,7 @@ type StakingEventWatcher struct {
 func NewStakingEventWatcher(
 	btcNotifier notifier.ChainNotifier,
 	btcClient btcclient.BTCClient,
+	indexer SpendChecker,
 	babylonNodeAdapter BabylonNodeAdapter,
 	cfg *config.BTCStakingTrackerConfig,
 	parentLogger *zap.Logger,
@@ -106,6 +108,7 @@ func NewStakingEventWatcher(
 		logger:                          parentLogger.With(zap.String("module", "staking_event_watcher")).Sugar(),
 		btcNotifier:                     btcNotifier,
 		btcClient:                       btcClient,
+		indexer:                         indexer,
 		babylonNodeAdapter:              babylonNodeAdapter,
 		metrics:                         metrics,
 		unbondingTracker:                NewTrackedDelegations(),
@@ -184,6 +187,10 @@ func (sew *StakingEventWatcher) handleNewBlocks(blockNotifier *notifier.BlockEpo
 			}
 			sew.currentBestBlockHeight.Store(uint32(block.Height))
 			sew.logger.Debugf("Received new best btc block: %d", block.Height)
+
+			if err := sew.checkSpend(); err != nil {
+				sew.logger.Errorf("error checking spend: %v", err)
+			}
 		case <-sew.quit:
 			return
 		}
@@ -418,19 +425,7 @@ func (sew *StakingEventWatcher) reportUnbondingToBabylon(
 	)
 }
 
-func (sew *StakingEventWatcher) watchForSpend(spendEvent *notifier.SpendEvent, td *TrackedDelegation) {
-	defer sew.wg.Done()
-	quitCtx, cancel := sew.quitContext()
-	defer cancel()
-
-	var spendingTx *wire.MsgTx
-	select {
-	case spendDetail := <-spendEvent.Spend:
-		spendingTx = spendDetail.SpendingTx
-	case <-sew.quit:
-		return
-	}
-
+func (sew *StakingEventWatcher) handleSpend(ctx context.Context, spendingTx *wire.MsgTx, td *TrackedDelegation) {
 	_, err := tryParseStakerSignatureFromSpentTx(spendingTx, td)
 	delegationID := td.StakingTx.TxHash()
 	spendingTxHash := spendingTx.TxHash()
@@ -442,25 +437,25 @@ func (sew *StakingEventWatcher) watchForSpend(spendEvent *notifier.SpendEvent, t
 		// As we only care about unbonding transactions, we do not need to take additional actions.
 		// We start polling babylon for delegation to stop being active, and then delete it from unbondingTracker.
 		sew.logger.Debugf("Spending tx %s for staking tx %s is not unbonding tx. Info: %v", spendingTxHash, delegationID, err)
-		proof := sew.waitForStakeSpendInclusionProof(quitCtx, spendingTx)
+		proof := sew.waitForStakeSpendInclusionProof(ctx, spendingTx)
 		if proof == nil {
 			sew.logger.Errorf("unbonding tx %s for staking tx %s proof not built", spendingTxHash, delegationID)
 
 			return
 		}
-		sew.reportUnbondingToBabylon(quitCtx, delegationID, spendingTx, proof)
+		sew.reportUnbondingToBabylon(ctx, delegationID, spendingTx, proof)
 	} else {
 		sew.metrics.DetectedUnbondingTransactionsCounter.Inc()
 		// We found valid unbonding tx. We need to try to report it to babylon.
 		// We stop reporting if delegation is no longer active or we succeed.
-		proof := sew.waitForStakeSpendInclusionProof(quitCtx, spendingTx)
+		proof := sew.waitForStakeSpendInclusionProof(ctx, spendingTx)
 		if proof == nil {
 			sew.logger.Errorf("unbonding tx %s for staking tx %s proof not built", spendingTxHash, delegationID)
 
 			return
 		}
 		sew.logger.Debugf("found unbonding tx %s for staking tx %s", spendingTxHash, delegationID)
-		sew.reportUnbondingToBabylon(quitCtx, delegationID, spendingTx, proof)
+		sew.reportUnbondingToBabylon(ctx, delegationID, spendingTx, proof)
 		sew.logger.Debugf("unbonding tx %s for staking tx %s reported to babylon", spendingTxHash, delegationID)
 	}
 
@@ -469,6 +464,66 @@ func (sew *StakingEventWatcher) watchForSpend(spendEvent *notifier.SpendEvent, t
 		&delegationInactive{stakingTxHash: delegationID},
 		sew.quit,
 	)
+}
+
+func (sew *StakingEventWatcher) checkSpend() error {
+	var wg sync.WaitGroup
+
+	for del := range sew.unbondingTracker.DelegationsIter(1000) {
+		if del.InProgress {
+			continue
+		}
+		ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
+		response, err := sew.indexer.GetOutspend(ctx, del.StakingTx.TxHash().String(), del.StakingOutputIdx)
+		cancel()
+
+		if err != nil {
+			sew.logger.Errorf("error getting outspend for staking tx %s: %v", del.StakingTx.TxHash(), err)
+
+			continue
+		}
+
+		if !response.Spent {
+			continue
+		}
+
+		if err := sew.unbondingTracker.UpdateActivation(del.StakingTx.TxHash(), true); err != nil {
+			return err
+		}
+
+		wg.Add(1)
+
+		// nolint:contextcheck
+		go func() {
+			defer wg.Done()
+			defer func() {
+				if err := sew.unbondingTracker.UpdateActivation(del.StakingTx.TxHash(), false); err != nil {
+					sew.logger.Warnf("error updating activation status for staking tx %s: %v", del.StakingTx.TxHash(), err)
+				}
+			}()
+			innerCtx, innerCancel := sew.quitContext()
+			defer innerCancel()
+
+			txHash, err := chainhash.NewHashFromStr(response.TxID)
+			if err != nil {
+				sew.logger.Errorf("error parsing tx hash %s: %v", response.TxID, err)
+
+				return
+			}
+			spendingTx, err := sew.btcClient.GetRawTransaction(txHash)
+			if err != nil {
+				sew.logger.Errorf("error getting spending tx %s: %v", response.TxID, err)
+
+				return
+			}
+
+			sew.handleSpend(innerCtx, spendingTx.MsgTx(), del)
+		}()
+	}
+
+	wg.Wait()
+
+	return nil
 }
 
 func (sew *StakingEventWatcher) buildSpendingTxProof(spendingTx *wire.MsgTx) (*btcstakingtypes.InclusionProof, error) {
@@ -537,7 +592,7 @@ func (sew *StakingEventWatcher) handleUnbondedDelegations() {
 		case activeDel := <-sew.unbondingDelegationChan:
 			sew.logger.Debugf("Received new delegation to watch for staking transaction with hash %s", activeDel.stakingTxHash)
 
-			del, err := sew.unbondingTracker.AddDelegation(
+			_, err := sew.unbondingTracker.AddDelegation(
 				activeDel.stakingTx,
 				activeDel.stakingOutputIdx,
 				activeDel.unbondingOutput,
@@ -552,26 +607,6 @@ func (sew *StakingEventWatcher) handleUnbondedDelegations() {
 			}
 
 			sew.metrics.NumberOfTrackedActiveDelegations.Inc()
-
-			stakingOutpoint := wire.OutPoint{
-				Hash:  activeDel.stakingTxHash,
-				Index: activeDel.stakingOutputIdx,
-			}
-
-			spendEv, err := sew.btcNotifier.RegisterSpendNtfn(
-				&stakingOutpoint,
-				activeDel.stakingTx.TxOut[activeDel.stakingOutputIdx].PkScript,
-				activeDel.delegationStartHeight,
-			)
-
-			if err != nil {
-				sew.logger.Errorf("error registering spend ntfn for staking tx %s: %v", activeDel.stakingTxHash, err)
-
-				continue
-			}
-
-			sew.wg.Add(1)
-			go sew.watchForSpend(spendEv, del)
 		case in := <-sew.unbondingRemovalChan:
 			sew.logger.Debugf("Delegation for staking transaction with hash %s stopped being active", in.stakingTxHash)
 			// remove delegation from unbondingTracker
@@ -616,7 +651,7 @@ func (sew *StakingEventWatcher) checkBtcForStakingTx() {
 	}
 
 	for del := range sew.pendingTracker.DelegationsIter(1000) {
-		if del.ActivationInProgress {
+		if del.InProgress {
 			continue
 		}
 		txHash := del.StakingTx.TxHash()