+++ CHANGELOG.md
@@ -40,6 +40,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Improvements
 
 * [#229](https://github.com/babylonlabs-io/vigilante/pull/229) chore: bigger batch size for fetching delegations
+* [#233](https://github.com/babylonlabs-io/vigilante/pull/233) chore: cleanup ckpt cache
 * [#232](https://github.com/babylonlabs-io/vigilante/pull/232) chore: comply to rbf policy
 
 ### Bug Fixes
@@ -47,8 +48,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * [#209](https://github.com/babylonlabs-io/vigilante/pull/209) fix: wait until slashing tx k-deep
 * [#223](https://github.com/babylonlabs-io/vigilante/pull/223) fix: consider minimal fee for bump
 * [#226](https://github.com/babylonlabs-io/vigilante/pull/226) fix: reselect inputs after adding manual output
+* [#229](https://github.com/babylonlabs-io/vigilante/pull/229) chore: lax unecessary btc tx checks
 
 ## v0.19.9
+
+### Improvements
+
+* [#211](https://github.com/babylonlabs-io/vigilante/pull/211) feat: adds indexer to btc staking tracker
 
 ### Bug Fixes